requirements.txt
# Podstawowy pakiet LangChain, zawiera BaseTool i inne rdzenne elementy
langchain
langchain-openai
python-dotenv
aiohttp